Restructure the code to have the event handling in the policy file
