+++ internal/factsengine/gatherers/facts.go
@@ -1,5 +1,12 @@
 package gatherers
 
+import (
+	"bytes"
+	"encoding/json"
+
+	"github.com/pkg/errors"
+)
+
 type FactsResult struct {
 	ExecutionID string `json:"execution_id"`
 	AgentID     string `json:"agent_id"`
@@ -35,4 +42,18 @@ func NewFactWithRequest(req FactRequest, value interface{}) Fact {
 		CheckID: req.CheckID,
 		Value:   value,
 	}
+}
+
+func PrettifyFactResult(fact Fact) (string, error) {
+	jsonResult, err := json.Marshal(fact)
+	if err != nil {
+		return "", errors.Wrap(err, "Error building the response")
+	}
+
+	var prettyJSON bytes.Buffer
+	if err := json.Indent(&prettyJSON, jsonResult, "", "  "); err != nil {
+		return "", errors.Wrap(err, "Error indenting the json data")
+	}
+
+	return prettyJSON.String(), nil
 }
